+++ iexec-core-library/src/main/java/com/iexec/core/api/SchedulerClient.java
@@ -35,6 +35,10 @@ import java.security.Signature;
 import java.util.List;
 
 public interface SchedulerClient {
+    /**
+     * @deprecated Use iexec-task-feedback-api
+     */
+    @Deprecated(forRemoval = true)
     @RequestLine("GET /metrics")
     PlatformMetric getMetrics();
 
@@ -85,18 +89,35 @@ public interface SchedulerClient {
     // endregion
 
     // region /tasks
+
+    /**
+     * @deprecated Use iexec-task-feedback-api
+     */
+    @Deprecated(forRemoval = true)
     @RequestLine("GET /tasks/{chainTaskId}")
     TaskModel getTask(@Param("chainTaskId") String chainTaskId);
 
+    /**
+     * @deprecated Use iexec-task-feedback-api
+     */
+    @Deprecated(forRemoval = true)
     @RequestLine("GET /tasks/logs/challenge?address={address}")
     EIP712Challenge getTaskLogsChallenge(@Param("address") String address);
 
+    /**
+     * @deprecated Use iexec-task-feedback-api
+     */
+    @Deprecated(forRemoval = true)
     @Headers("Authorization: {authorization}")
     @RequestLine("GET /tasks/{chainTaskId}/logs")
     TaskLogsModel getTaskLogs(
             @Param("chainTaskId") String chainTaskId,
             @Param("authorization") String authorization);
 
+    /**
+     * @deprecated Use iexec-task-feedback-api
+     */
+    @Deprecated(forRemoval = true)
     @Headers("Authorization: {authorization}")
     @RequestLine("GET /tasks/{chainTaskId}/replicates/{walletAddress}/logs")
     ComputeLogs getComputeLogs(