+++ iexec-core-library/src/main/java/com/iexec/core/task/TaskModel.java
@@ -23,6 +23,10 @@ import lombok.*;
 import java.util.Date;
 import java.util.List;
 
+/**
+ * @deprecated Use iexec-task-feedback-api
+ */
+@Deprecated(forRemoval = true)
 @Getter
 @Setter
 @NoArgsConstructor