+++ iexec-task-feedback-api/src/main/java/com/iexec/core/replicate/ReplicateModel.java
@@ -22,10 +22,6 @@ import lombok.*;
 
 import java.util.List;
 
-/**
- * @deprecated Use iexec-task-feedback-api
- */
-@Deprecated(forRemoval = true)
 @Getter
 @Setter
 @Builder