+++ iexec-task-feedback-api/src/main/java/com/iexec/core/replicate/ReplicateStatusUpdateModel.java
@@ -28,10 +28,6 @@ import lombok.NoArgsConstructor;
 
 import java.util.Date;
 
-/**
- * @deprecated Use iexec-task-feedback-api
- */
-@Deprecated(forRemoval = true)
 @Data
 @Builder
 @NoArgsConstructor